v1.3

+++ controllers/pets/deletePet.js
@@ -3,11 +3,19 @@ const { HttpError } = require("../../helpers");
 
 const deletePet = async (req, res, next) => {
   const id = req.params.id;
-  const removedContact = await Pet.findByIdAndDelete(id);
-  if (!removedContact) {
-    throw HttpError(404, "Contact not found");
+  console.log(req.user);
+  const ownerId = req.user._id;
+  const pet = await Pet.findById(id);
+  if (!pet) {
+    throw HttpError(404, "Pet not found");
   }
-  res.status(204).json({ message: "Contact deleted" });
+  console.log(pet.owner, "OwnerId", ownerId, "from auth");
+  if (pet.owner !== ownerId) {
+    throw HttpError(401, "not authorized to delete");
+  }
+  await Pet.findByIdAndDelete(id);
+
+  res.status(204).json({ message: "Pet deleted" });
 };
 
 module.exports = deletePet;
